+++ pkg/message/adaptor/all.go
@@ -1,7 +1,6 @@
 package adaptor
 
 import (
-	_ "github.com/compose/transporter/pkg/message/adaptor/elasticsearch"
 	_ "github.com/compose/transporter/pkg/message/adaptor/file"
 	// _ "github.com/compose/transporter/pkg/message/adaptor/influxdb"
 	_ "github.com/compose/transporter/pkg/message/adaptor/rethinkdb"